Separate exporter test classes
